app/backend-v1/file-processor/requirements.txt
boto3==1.20.42
fastapi
pymysql
requests
uvicorn
python-dotenv